client/section/board/components/write/posting.js의 미리보기 카드를 별도 컴포넌트인 previewCard.js로 분리했습니다.
